Move to unique files and add temp callbacks
